lessons/django-oauth/requirements.txt
# Dependências
Django==5.2
channels==4.2.2
daphne==4.1.2
redis==6.0.0
djangorestframework==3.14.0
django-oauth-toolkit==3.0.1

# Variáveis de ambiente arquivo .env
python-decouple==3.8
